pyproject.toml
[project]
name = "ai-canvas-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.0",
    "motor>=3.7.1",
    "ollama>=0.6.1",
    "pymongo>=4.16.0",
    "pypdf>=6.7.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
